Initial Commit - broken out from VPS
This is the single page Vimwiki 2 HTML Markdown converter that was
started as part of the vimwiki publishing system (which is bound to be
renamed).  This is just an extraction of it for the simple purposes of
using it with the :Vimwiki2HTML command(s).  It understands Vimwiki
[[links]] (but on this simple ones.  It doesn't do anchors or intersite
links, and adds .html whether it is needed or not.  I may fix these
issues as time goes on.

Anyway a first commit.